Finish the logic and update login redirects
