+++ www/src/components/users/MagicLogin.js
@@ -10,14 +10,11 @@ import {
 import {
   useApolloClient, useLazyQuery, useMutation, useQuery,
 } from '@apollo/client'
-import {
-  Navigate, useLocation, useNavigate, useParams,
-} from 'react-router-dom'
+import { Navigate, useLocation, useParams } from 'react-router-dom'
 import queryString from 'query-string'
 import {
   A, Article, Button, Div, Flex, H2, Icon, Img, Input, P,
 } from 'honorable'
-
 import { WelcomeHeader } from 'components/utils/WelcomeHeader'
 
 import { fetchToken, setToken } from '../../helpers/authentication'
@@ -26,8 +23,7 @@ import { disableState } from '../Login'
 import { PLURAL_FULL_LOGO_WHITE, PLURAL_MARK_WHITE } from '../constants'
 import { ACCEPT_LOGIN } from '../oidc/queries'
 import { host } from '../../helpers/hostname'
-
-import { useHistory } from '../../history/router'
+import { useHistory } from '../../router'
 
 import {
   METHOD_ICONS, getDeviceToken, saveChallenge, saveDeviceToken, wipeChallenge, wipeDeviceToken,
@@ -240,7 +236,7 @@ export function handleOauthChallenge(client, challenge) {
 }
 
 function LoginPoller({ challenge, token, deviceToken }) {
-  const navigate = useNavigate()
+  const history = useHistory()
   const client = useApolloClient()
   const [success, setSuccess] = useState(false)
 
@@ -259,13 +255,13 @@ function LoginPoller({ challenge, token, deviceToken }) {
           handleOauthChallenge(client, challenge)
         }
         else {
-          navigate('/')
+          history.navigate('/')
         }
       })
     }, 2000)
 
     return () => clearInterval(interval)
-  }, [token, challenge, deviceToken, navigate, client])
+  }, [token, challenge, deviceToken, history, client])
 
   if (success) {
     return (
@@ -287,7 +283,6 @@ function LoginPoller({ challenge, token, deviceToken }) {
 }
 
 export function Login() {
-  const navigate = useNavigate()
   const history = useHistory()
   const client = useApolloClient()
   const location = useLocation()
@@ -298,8 +293,6 @@ export function Login() {
   const [getLoginMethod, { data, loading: qLoading, error: qError }] = useLazyQuery(LOGIN_METHOD, {
     variables: { email, host: host() },
   })
-
-  console.log(history)
 
   const loginMethod = data?.loginMethod?.loginMethod
   const open = loginMethod === LoginMethod.PASSWORD
@@ -316,7 +309,7 @@ export function Login() {
         handleOauthChallenge(client, challenge)
       }
       else {
-        navigate('/')
+        history.navigate('/')
       }
     },
   })
@@ -338,9 +331,9 @@ export function Login() {
       handleOauthChallenge(client, challenge)
     }
     else if (!deviceToken && jwt) {
-      navigate('/')
+      history.navigate('/')
     }
-  }, [challenge, deviceToken, navigate, client, jwt])
+  }, [challenge, deviceToken, history, client, jwt])
 
   const submit = useCallback(() => (open ? mutation() : getLoginMethod()), [mutation, getLoginMethod, open])
 
@@ -412,7 +405,7 @@ export function Login() {
                   caption={(
                     <A
                       inline
-                      onClick={() => navigate('/password-reset')}
+                      onClick={() => history.navigate('/password-reset')}
                     >forgot your password?
                     </A>
                   )}
@@ -490,7 +483,7 @@ function OAuthOption({ url: { authorizeUrl, provider }, ...props }) {
 }
 
 export function Signup() {
-  const navigate = useNavigate()
+  const history = useHistory()
   const location = useLocation()
   const [email, setEmail] = useState(location?.state?.email || '')
   const [password, setPassword] = useState('')
@@ -503,16 +496,16 @@ export function Signup() {
     onCompleted: ({ signup: { jwt } }) => {
       if (deviceToken) finishedDeviceLogin()
       setToken(jwt)
-      window.location = '/shell'
+      history.navigate('/shell')
     },
   })
   const { data } = useQuery(OAUTH_URLS, { variables: { host: host() } })
 
   useEffect(() => {
     if (fetchToken()) {
-      navigate('/')
+      history.navigate('/')
     }
-  }, [navigate])
+  }, [history])
 
   const { disabled, reason } = disableState(password, confirm, email)
 
@@ -617,7 +610,7 @@ export function Signup() {
         Already have an account?{' '}
         <A
           inline
-          onClick={() => navigate('/login')}
+          onClick={() => history.navigate('/login')}
         >
           Login
         </A>
