+++ www/src/router/browser.ts
@@ -2,4 +2,4 @@ import { createBrowserHistory } from 'history'
 
 const browserHistory = createBrowserHistory()
 
-export default browserHistory
+export { browserHistory }
