+++ www/src/router/index.ts
@@ -1,0 +1,5 @@
+import { useHistory } from './hook'
+import { HistoryRouter } from './component'
+import { browserHistory } from './browser'
+
+export { HistoryRouter, useHistory, browserHistory }
